requirements.txt
# Core
requests>=2.31.0
boto3>=1.34.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.2.0
jsonschema>=4.21.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
moto[s3]>=5.0.0

# Docs (openapi.yaml → swagger.json): python docs/generate_swagger_json.py
PyYAML>=6.0.1